conditional install flyway

+++ roles/lobby_db/tasks/install_flyway.sh
@@ -7,11 +7,11 @@ URL="https://repo1.maven.org/maven2/org/flywaydb/flyway-commandline/5.0.7/flyway
 
 FLYWAY_FOLDER="/home/triplea/flyway-5.0.7"
 
-wget $URL
-tar xvf flyway*tar.gz
-rm *tar.gz
+if [ ! -d ${FLYWAY_FOLDER} ]; then
+  wget $URL
+  tar xvf flyway*tar.gz
+  rm *tar.gz
 
-
-mv flyway* ${FLYWAY_FOLDER}
-chown -R triplea:triplea /home/triplea/
-
+  mv flyway* ${FLYWAY_FOLDER}
+  chown -R triplea:triplea /home/triplea/
+fi
